[all] signedness, narrowing conversion fixes

+++ dune/stuff/grid/output/pgf.hh
@@ -38,7 +38,7 @@ struct PgfCoordWrapper : FieldVector<double, 2>
   PgfCoordWrapper(const FieldVector<double, wdim>& vector)
     : FieldVector<double, 2>(0.0)
   {
-    for (size_t i = 0; i < std::min(int(wdim), 2); ++i)
+    for (size_t i = 0; i < size_t(std::min(wdim, 2)); ++i)
       (*this)[i] = vector[i];
   }
 };
